dual-momentum-mvp/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.32.*
requests==2.32.*
python-dateutil==2.9.*
python-dotenv==1.0.*
